Initial commit
This was originally built for a game being developed by @billythekid
and I. I’ve stripped out any game specific things. I have left in
placeholder functions for connecting to a server; this can be used to
extend commands, and provide tamper proof online functionality.

Feel free to use this in your project, but please follow the license.