Mostly adding comments & fixed BOLD bug.
Added a number of comments (mostly for my own benefit as I'm new to js).
Also added a case for tab complete (Line 63)- though this may require a bit of a fiddle to implement.
And fixed a bug! Where typing open triangle brackets and any character "<a" previously resulted in the <bold/> not working. Basically old code had: <bold>"<a"<bold/>.....waiting for />. (Line 78 modified to fix)